btrfs: fix bitmap leak when loading free space cache on duplicate entry
[ Upstream commit 320d8dc612660da84c3b70a28658bb38069e5a9a ]

If we failed to link a free space entry because there's already a
conflicting entry for the same offset, we free the free space entry but
we don't free the associated bitmap that we had just allocated before.
Fix that by freeing the bitmap before freeing the entry.

Reviewed-by: Johannes Thumshirn <johannes.thumshirn@wdc.com>
Signed-off-by: Filipe Manana <fdmanana@suse.com>
Reviewed-by: David Sterba <dsterba@suse.com>
Signed-off-by: David Sterba <dsterba@suse.com>
Signed-off-by: Sasha Levin <sashal@kernel.org>
(cherry picked from commit fad0bb34cfcea693903409356693988f04715b8e)
[Vegard: use kfree() due to missing commit
 4874c6fe1c9efe704bf155afab268ead7c364c9b ("btrfs: fix allocation of
 free space cache v1 bitmap pages").]
Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>
Signed-off-by: firebird11 <hbgassel@gmail.com>